backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.27.0
